user-acount-service/requirements/base.txt
django==4.2.1
django-environ==0.10.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-filter==23.2
django-countries==7.5.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.13
drf-yasg==1.21.5
Pillow==9.5.0
argon2-cffi==21.3.0
pytz==2023.3
redis==4.5.5
django-allauth==0.54.0
djangorestframework-simplejwt==5.2.2
dj-rest-auth==4.0.1
django-elasticsearch-dsl==7.3.0
django-elasticsearch-dsl-drf==0.22.5
